Split `StringAssertionSpecs` file
